backend/requirements.txt
django==5.0.3
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
ultralytics==8.1.0
opencv-python==4.9.0.80
firebase-admin==6.4.0
python-multipart==0.0.9
